requirements.txt
PyYAML
telethon